refactor(admin): convert router to ts, extract orchestration into service
Finish Phase 3: the admin router was the last .js in the module (alongside the
already-ts service/diagnose files). Move the inline mongo aggregations
(datasets/applications errors, owners, base-applications listings) into
service.ts as express-free functions taking the parsed query, leaving thin
router handlers. Kills the ParsedQs->Record<string,string> errors by typing the
service params; admin 11 -> 1 tsc errors (the remaining one is the pervasive
publicBaseUrl baseline). listDatasetsWithEsWarnings keeps its (size, skip)
signature (reused by the diagnose service); a thin findDatasetsEsWarnings wrapper
adapts the route.

Ratchet 1757 -> 1748.
