+++ dev/type-errors-baseline.txt
@@ -1,1 +1,1 @@
-1757
+1748
